[Unittests] Add equal and hash to unit tests
I've implemented a bare bones notion of Equality. This is useful for
learning unittests, but I don't know if it is appropriate to have any
further notion of equality in neural networks, and it may in fact lead
to more confusion down the road. I am adding this commit with the
acknoledgement that implementing a notion of equality here may be
harmful, but is just to practice with Python's features.